made changes to messages provider
