+++ src/components/ApplicationViews.js
@@ -11,6 +11,10 @@ import { EventDetail } from "./Events/EventDetail"
 import { FriendProvider } from "./Friends/FriendProvider"
 import { FriendList } from "./Friends/FriendList"
 import { FriendForm } from "./Friends/FriendForm"
+import { MessageProvider } from "./Messages/MessageProvider"
+import { MessageList } from "./Messages/MessageList"
+import { MessageForm } from "./Messages/MessageForm"
+import { MessageDetail } from "./Messages/MessageDetail"
 
 export const ApplicationViews = () => {
   return (
@@ -18,6 +22,7 @@ export const ApplicationViews = () => {
       <EventProvider>
       <ArticleProvider>
       <FriendProvider>
+      <MessageProvider>
        
       <Route exact path="/">
         <ArticleList />
@@ -39,9 +44,16 @@ export const ApplicationViews = () => {
         <FriendForm />
       </Route>
     
-      <Route path="/messages">
-        {/* Render the component for the messages */}
+      <Route exact path="/messages">
+        <MessageList />
       </Route>
+      <Route path="/messages/create">
+        <MessageForm />
+      </Route>  
+      <Route path="/messages/detail/:messageId(\d+)">
+        <MessageDetail />
+      </Route>
+
       <Route exact path="/tasks">
         {/* Render the component for the user's tasks */}
       </Route>
@@ -57,6 +69,7 @@ export const ApplicationViews = () => {
       <Route exact path="/events/detail/:eventId(\d+)">
         <EventDetail />
       </Route>
+      </MessageProvider>
       </FriendProvider>
       </ArticleProvider>
       </EventProvider>
