+++ src/components/Messages/MessageDetail.js
@@ -26,7 +26,7 @@ export const MessageDetail = () => {
     return (
         <section className="message">
         <h3 className="message__body">{message.body}</h3>
-        <div className="message__user">user: {message.user?.name}</div>
+        <div className="message__user">user: {message.userId?.name}</div>
         <button onClick={deleteMessages}>Delete Message</button>
         </section>
     )
