+++ src/components/Messages/MessageForm.js
@@ -1,37 +1,40 @@
-import React, { useContext, useEffect, useState } from "react"
+import React, { useContext, useState, useEffect } from "react"
 import { useHistory } from "react-router-dom"
 import { MessageContext } from "./MessageProvider"
-import { UserContext } from "../user/UserProvider"
+
 
 export const MessageForm = () => {
-    const { addMessage } = useContext(MessageContext)
-    const { users, getUsers } = useContext(UserContext)
+    const { addMessage, messages, getMessages } = useContext(MessageContext)
     const history = useHistory()
-
+    
     const [message, setMessage] = useState({
-            body: "",
-            userId: 0,
-            timeStamp: 0,
+        userId: 0,
+        body: "",
+        timeStamp: 0,
     });
-
-
+    
+    useEffect(() => {
+        getMessages()
+    }, [])
+    
     const handleControlledInputChange = (event) => {
-        const newMessage = { ...message }
+        const newMessage = { ...messages }
         newMessage[event.target.id] = event.target.value
         setMessage(newMessage)
     }
-
+    
     const saveMessage = (event) => {
         event.preventDefault()
+            
+        const userId = (message.userId)
 
-        const userId = parseInt(message.userId)
-        
-        const newMessage = {
-            body: message.body,
-            userId: userId,
-            timeStamp: message.timeStamp,
+            const newMessage = {
+                body: message.body,
+                userId: userId,
+                timeStamp: message.timeStamp,
         }
-        addMessage(newMessage).then(() => history.push("/messages"))
+        addMessage(newMessage)
+            .then(() => history.push("/messages"))
         
     }
 
@@ -41,8 +44,8 @@ export const MessageForm = () => {
         <h2 className="messageForm_header">New Message</h2>
         <fieldset>
             <div className="form-group">
-            <label htmlFor="textBody">Please write your message</label>
-            <input type="text" id="textArea" required autoFocus className="form-control" placeholder="Message body" value={message.body} onChange={handleControlledInputChange} />
+            <label htmlFor="textBody">Please write your message below</label>
+            <input type="text" id="textBody" required autoFocus className="form-control" placeholder="New message" value={message.body} onChange={handleControlledInputChange} />
             </div>
         </fieldset>
         <fieldset>
